# guild-app/src/app/api/v1/auth/verify/route.ts
import { NextRequest, NextResponse } from "next/server"
import { verifyAndConsume } from "@/lib/rola"
import { createSession, setSessionCookie } from "@/lib/auth"
import { findOrCreateUser } from "@/db/queries/users"
import { verifyAuthSchema } from "@/lib/validation"
import { fromError } from "@/lib/api-response"
import { createRateLimiter, getClientIp, rateLimitResponse } from "@/lib/rate-limit"

const limiter = createRateLimiter({ windowMs: 60_000, max: 10 })

export async function POST(req: NextRequest) {
  try {
    const limit = limiter(getClientIp(req))
    if (!limit.ok) return rateLimitResponse(limit.retryAfter!)

    const body = await req.json().catch(() => null)
    if (!body) {
      return NextResponse.json(
        { ok: false, error: { code: "INVALID_BODY", message: "Invalid JSON" } },
        { status: 400 },
      )
    }

    const parsed = verifyAuthSchema.safeParse(body)
    if (!parsed.success) {
      return NextResponse.json(
        { ok: false, error: { code: "VALIDATION_ERROR", message: parsed.error.message } },
        { status: 400 },
      )
    }

    const { signed_challenge } = parsed.data
    const result = await verifyAndConsume(signed_challenge)
    if (!result.ok) {
      return NextResponse.json(
        { ok: false, error: { code: "AUTH_FAILED", message: result.error } },
        { status: 401 },
      )
    }

    // onConflictDoNothing (findOrCreateUser) never touches an existing row's
    // columns, so a returning user is never silently un-suspended here — the
    // only way suspendedAt clears is the operator's --lift (see
    // scripts/suspend-account.mjs). Refuse to mint a session for a suspended
    // address rather than let it sign in and hit ACCOUNT_SUSPENDED on its
    // first real call.
    const user = await findOrCreateUser(result.address)
    if (user.suspendedAt) {
      return NextResponse.json(
        {
          ok: false,
          error: {
            code: "ACCOUNT_SUSPENDED",
            message: "This account has been suspended. Contact the operator for details.",
          },
        },
        { status: 403 },
      )
    }
    const token = await createSession(user.id)
    await setSessionCookie(token)

    return NextResponse.json({ ok: true, data: { user } })
  } catch (err) {
    return fromError(err)
  }
}